WilsonLine
We want to compute a Wilson line starting from every point and following the same (relative) path along the lattice.

We start out with the unit matrix at every site and gauge transport it around until it lands at (original site) + (path displacement).

The commented out section of test.cc was used to test the gauge_transport function, which I've since hidden away in WilsonLine.cc.  It might make sense to expose it and stick it in its own header.