Sort out link orientation issue.
I previously thought the interpretation of the chroma plaquette calculation was that U[µ](x) contains the links that points TO x from x+µ, rather than FROM x to x+µ.  This was mistaken, but I had another, compensating mistake, which was that I had a missing complex conjugate in the parallel transport equation.  This explains why I was able to correctly compute the plaquette---it is real.

This commit adds in the complex conjugate, and updates the corresponding test of long-distance gauge transport.  (Note, again, that the test of the two-step plaquette didn't change---it previously got the right answer the wrong way!)